LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.direction.valid-expected.txt

2d.text.direction.valid
Actual output:

FAIL Canvas test: 2d.text.direction.valid assert_equals: ctx.direction === 'inherit' (got ltr[string], expected inherit[string]) expected "inherit" but got "ltr"
